Decimal based credit migration

+++ cron/jobs/auto_refunds.go
@@ -20,7 +20,7 @@ func (j *JobRunner) RefundOldGenerationCredits(log Logger) error {
 			db := tx.Client()
 			j.Repo.SetGenerationFailed(gen.ID.String(), shared.TIMEOUT_ERROR, 0, db)
 			// Upscale is always 1 credit
-			_, err := j.Repo.RefundCreditsToUser(gen.UserID, gen.NumOutputs, db)
+			_, err := j.Repo.RefundCreditsToUser(gen.UserID, float32(gen.NumOutputs), db)
 			if err != nil {
 				log.Errorf("Error refunding credits for generation %s %s %v", gen.UserID.String(), gen.ID.String(), err)
 				return err
